genai_chatbot_hackathon/requirements.txt
Flask
openai
pinecone
streamlit
azure-search-documents
python-dotenv